Use address argument instead of full-connection if address is defined
Since the beginning I wasn't glad with the full-connection argument name. I believe just address has the same meaning and it is simpler. My intention isn't to deprecate full-connection to not cause a breaking change for anyone. But address will be the default argument for the address information and full-connection will only be used if address isn't defined.